+++ app/src/main/java/com/vmmontes/mytest/data/dataSource/VehicleDataSourceContract.kt
@@ -1,0 +1,7 @@
+package com.vmmontes.mytest.data.dataSource
+
+import com.vmmontes.mytest.domain.model.VehicleDomainModel
+
+interface VehicleDataSourceContract {
+    fun getVehicles(): List<VehicleDomainModel>
+}
