+++ app/src/main/java/com/vmmontes/mytest/data/dataSource/VehiclesDataSource.kt
@@ -1,15 +1,13 @@
-package com.vmmontes.mytest.dataSource
+package com.vmmontes.mytest.data.dataSource
 
 import com.api.MyTaxpiApiContract
-import com.api.retrofit.MyTaxiApiClient
-import com.vmmontes.mytest.dataSource.mapper.toDomain
+import com.vmmontes.mytest.data.mapper.toDomain
 import com.vmmontes.mytest.domain.model.VehicleDomainModel
 
-class ApiData {
+class VehiclesDataSource constructor(
+    private val taxiApiClient: MyTaxpiApiContract) : VehicleDataSourceContract {
 
-    val taxiApiClient: MyTaxpiApiContract = MyTaxiApiClient()
-
-    fun getVehicles(): List<VehicleDomainModel> {
+    override fun getVehicles(): List<VehicleDomainModel> {
         val vehiclesApiModel = taxiApiClient.getVehicles()
 
         val vehiclesDomainModel = mutableListOf<VehicleDomainModel>()
@@ -20,5 +18,4 @@ class ApiData {
 
         return vehiclesDomainModel
     }
-
 }
